some more splitting
