+++ decompiler/pipeline/controlflowanalysis/restructuring_commons/region_finder/__init__.py
@@ -1,0 +1,2 @@
+from .cyclic_region_finder import CyclicRegionFinder
+from .acyclic_region_finder import AcyclicRegionFinder
